requirements.txt
Django==3.2.9
django-background-tasks==1.2.5
django-celery==3.3.1
django-compat==1.0.15
django-crispy-forms==1.13.0
django-debug-permissions==1.0.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
drf-yasg==1.20.0
psycopg2==2.9.1
openapi-codec==1.3.2
gunicorn==20.0.4
whitenoise==5.3.0
django-allauth==0.45.0
channels==3.0.4
